LayoutTests/imported/w3c/web-platform-tests/html/browsers/history/the-history-interface/007-expected.txt

It looks like the browser stopped loading the page when encountering a .go(-1) command pointing to a pushed state. This will break the tests.

Harness Error (TIMEOUT), message = null

PASS history.state should initially be null
PASS history.pushState support is needed for this testcase
PASS history.state should reflect pushed state
